Move default implementation of AstSchemaExtender ot its own class
